Attempt to not show the same ad back to back
Added a new primitive, `VarietyQueue` which which can have values `push`ed to
it, and for each `pop`d value, it will make a localized best-effort to return
something it hasn't returned recently.

Building from that is the `VarietyStream` which once told how to generate new
values and compare equality, will feed semi-unique values into a piped stream.

Finally, added a `VariedAdStream` which works very similar to the current
`AdStream`, but uses the tools above. After watching it play for half an hour or
so, it appears to now show duplicates.

`AdCache` has been taken out of the pipeline for the reasons outlined below

Notes:
- The interaction between the AdStream and AdCache was downloading far more
  advertisements than the `queueSize` config value.
- `AdCache`, for reasons still to be determined, was re-ordering ads that passed
  through it.